pyproject.toml
[project]
name = "rag-assistant"
version = "0.1.0"
description = "Rag assistant for your documents"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.4",
    "llama-index-core>=0.14.7",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.7",
    "llama-index-vector-stores-chroma>=0.5.3",
    "openai>=1.109.1",
    "python-dotenv>=1.2.1",
]
